simplifying events sorting before creating time tokens, fixed resampling tpb
